#114 Phase A: extract Lua res:* helpers -> bindings_response.c
The core lua_rt_bindings.o referenced Keel's response API (kl_response_status/
header/body_*, hl_maybe_compress) purely via the res:* handler helpers
(res:status/header/json/html/text/bytes/redirect + the response metatable +
hl_lua_make_response). Move that whole cluster into a new bindings_response.c so
the core bindings object holds ZERO Keel-response references - they now live
only on the HTTP side of the seam.

- bindings_response.c: the res:* helpers + response metatable + the public
  hl_lua_make_response (dispatch.c still calls it, resolved from the new object).
- bindings.c: de-static get_hl_lua_from_L (shared with the moved cluster) and
  drop the res:* block + the HL_RESPONSE_MT define.
- internal.h: declare get_hl_lua_from_L.

Auto-wired into the lua runtime archive via LUA_RT_OBJS; Phase C relocates it
into the composed `http` feature alongside http_register.c.

Verified: `make test` 60/60; e2e_feature_runtime green (the lua res:json app
serves); `nm` confirms lua_rt_bindings.o has no kl_response_* / hl_maybe_compress
undefined refs - they're all in lua_rt_bindings_response.o.
